backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
python-dotenv==1.0.0
openai==1.45.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx<0.28

